requirements.txt
attrs==21.2.0
Django==2.2.10
djangorestframework==3.12.4
generics==3.4.0
importlib-metadata==4.7.1
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
views==0.3
zipp==3.5.0
